Create a private property to store names and a method to return them.
It was easy right?, let's review how we can test this method. We need to think about how we can test the private property. Our instinct is to think about why is necesary to put name privates and why not just simply put default accessor or public instead. However there are a Javascript trick (only javascript), by accesing the property in this way servive['privateProperty']. So, now, we know it is important be aware about public and private properties and methods have a different strategy. Let's review then how the test is implemented.