Writing the minimum only defined unit test for conferencesService.
As you may see the service looks more or less similiar to our names provider, one of the main differences is the HttpClient giving us the way to do a get request in a pretty straigforward way, similarely we can do get, post, put, delete etc, it is interesting because you can do as much as you need with HTTP Client, the natural return of the HTTP Client is an observable, and right now maybe you can't find too much profit of it, however later on we will cover it. Let's review then how to test this simple service first let's try with the typical structure of the test, the goal is only to test whether the service is defined or not.