Solution about the excercise 4 testing the post request.
What is interesting about this is the response contains a key value which is the id and it should be the key value here to test if everything is going ok or not. Nothing should be very different than the get endpoint except we are using the interface and we are asserting more data focused.