Mocking the service to handle the authentication process, changing some flows
Now, We have put the service to handle the authentication process, but it is not the most important part about this. Really we shouldn't care about because, first from our test we can handle and determine what are we going to show if the authentication process is success or not. Second we need to learn the first way to mock the service from the component and redefine some small flow about. Be carefull with the partial the providers.userValue and how reestructure the flow.