Setting the final exercise.
It has been a long journey until here, so take this exercise as your final exam. You should do two things, first is to write the test for the authenticationService which you will find already implemented using a hardcoded validation. Second, you will notice about a variable in the loginComponent called "isLogged" this variable handle a conditional in the view to display the protected content in the view. So your goal is to write the tests for this functionality.